created player response struct

+++ src/main/game.go
@@ -12,8 +12,8 @@ func (g Game) AddPlayer(name string) Player {
     player := Player {
         numPlayers,
         name,
-        rand.Float32() * 4000,
-        rand.Float32() * 4000,
+        rand.Float64() * 4000,
+        rand.Float64() * 4000,
     }
     numPlayers++
     g.Players = append(g.Players, player)
